added classes for each document processing step
